Edited script to restart waybar

+++ .config/waybar/waybar.sh
@@ -4,8 +4,16 @@ run_waybar() {
     waybar -c ~/.config/waybar/config  -s ~/.config/waybar/style.css
 }
 
+waybar_active() {
+    pid=$(pidof waybar)a
+    if [ -z "$pid" ]; then
+        return 1
+    fi
+    return 0
+}
+
 restart_waybar() {
-    pid=$(pgrep waybar)
+  if waybar_active; then
     for i in $pid; do
         kill $i
     done
